moved all fragments and activites to ui folder + deleted old classes
