+++ app/src/androidTest/java/com/github/sdp/mediato/AuthenticationActivityTest.java
@@ -12,7 +12,7 @@ import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static androidx.test.platform.app.InstrumentationRegistry.getInstrumentation;
-import static com.github.sdp.mediato.AuthenticationActivity.isNetworkAvailable;
+import static com.github.sdp.mediato.ui.AuthenticationActivity.isNetworkAvailable;
 import static org.hamcrest.Matchers.not;
 import static org.junit.Assert.assertTrue;
 import static java.lang.Thread.sleep;
@@ -33,6 +33,9 @@ import com.github.sdp.mediato.DatabaseTests.DataBaseTestUtil;
 import com.github.sdp.mediato.data.UserDatabase;
 import com.github.sdp.mediato.model.Location;
 import com.github.sdp.mediato.model.User;
+import com.github.sdp.mediato.ui.AuthenticationActivity;
+import com.github.sdp.mediato.ui.MainActivity;
+import com.github.sdp.mediato.ui.NewProfileActivity;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.auth.AuthResult;
