+++ app/src/androidTest/java/com/github/sdp/mediato/MyFollowersFragmentTest.java
@@ -15,8 +15,8 @@ import com.github.sdp.mediato.DatabaseTests.DataBaseTestUtil;
 import com.github.sdp.mediato.data.UserDatabase;
 import com.github.sdp.mediato.model.Location;
 import com.github.sdp.mediato.model.User;
+import com.github.sdp.mediato.ui.MainActivity;
 import com.github.sdp.mediato.ui.MyFollowersFragment;
-import com.github.sdp.mediato.ui.MyFollowingFragment;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
