+++ app/src/androidTest/java/com/github/sdp/mediato/NewItemFragmentTest.java
@@ -13,8 +13,6 @@ import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static androidx.test.platform.app.InstrumentationRegistry.getInstrumentation;
 import static com.adevinta.android.barista.interaction.BaristaClickInteractions.clickOn;
-import static com.adevinta.android.barista.interaction.BaristaEditTextInteractions.typeTo;
-import static com.adevinta.android.barista.interaction.BaristaKeyboardInteractions.closeKeyboard;
 import static com.github.sdp.mediato.ui.NewItemFragment.MAX_REVIEW_LENGTH;
 
 import android.app.Activity;
@@ -38,6 +36,7 @@ import androidx.test.uiautomator.UiDevice;
 import com.github.sdp.mediato.model.Review;
 import com.github.sdp.mediato.model.media.Media;
 import com.github.sdp.mediato.model.media.Movie;
+import com.github.sdp.mediato.ui.MainActivity;
 import com.github.sdp.mediato.ui.NewItemFragment;
 
 import org.hamcrest.Description;
@@ -46,8 +45,6 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-
-import java.util.Locale;
 
 /**
  * Test class for new item activity, used to add ratings and comments
