+++ app/src/main/java/com/github/sdp/mediato/ui/MyFollowersFragment.java
@@ -11,7 +11,6 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.github.sdp.mediato.MainActivity;
 import com.github.sdp.mediato.R;
 import com.github.sdp.mediato.ui.viewmodel.MyFollowersViewModel;
 import com.github.sdp.mediato.utility.adapters.UserAdapter;
