+++ app/src/main/java/com/github/sdp/mediato/ui/MyFollowingFragment.java
@@ -14,7 +14,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.github.sdp.mediato.MainActivity;
 import com.github.sdp.mediato.R;
 import com.github.sdp.mediato.utility.adapters.UserAdapter;
 import com.github.sdp.mediato.ui.viewmodel.MyFollowingViewModel;
