+++ app/src/main/java/com/github/sdp/mediato/ui/NewProfileActivity.java
@@ -1,9 +1,10 @@
-package com.github.sdp.mediato;
+package com.github.sdp.mediato.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
+import com.github.sdp.mediato.R;
 
 
 /**
