+++ app/src/main/java/com/github/sdp/mediato/ui/ReadOnlyProfileFragment.java
@@ -1,7 +1,6 @@
 package com.github.sdp.mediato.ui;
 
 import android.os.Bundle;
-import android.os.Handler;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -9,18 +8,10 @@ import android.widget.Button;
 
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-import com.github.sdp.mediato.MainActivity;
 import com.github.sdp.mediato.R;
-import com.github.sdp.mediato.data.CollectionsDatabase;
-import com.github.sdp.mediato.data.UserDatabase;
 import com.github.sdp.mediato.model.media.Collection;
-import com.github.sdp.mediato.ui.viewmodel.ReadOnlyProfileViewModel;
-import com.github.sdp.mediato.utility.SampleReviews;
 import com.github.sdp.mediato.utility.adapters.CollectionListAdapter;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import org.checkerframework.checker.units.qual.C;
 
 /**
  * A fragment to display another user's profile. It extends the basic profile fragment to also include:
