+++ app/src/main/java/com/github/sdp/mediato/utility/adapters/MediaAdapter.java
@@ -10,10 +10,9 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
-import com.github.sdp.mediato.FragmentSwitcher;
+import com.github.sdp.mediato.utility.FragmentSwitcher;
 import com.github.sdp.mediato.R;
 import com.github.sdp.mediato.databinding.AdapterMediaItemBinding;
-import com.github.sdp.mediato.databinding.LayoutMovieItemBinding;
 import com.github.sdp.mediato.model.media.Media;
 import com.github.sdp.mediato.ui.NewItemFragment;
