+++ app/src/main/res/layout/activity_authentication.xml
@@ -5,7 +5,7 @@
     android:id="@+id/authentication_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".ui.MainActivity">
 
     <com.google.android.gms.common.SignInButton
         android:id="@+id/google_sign_in"
