+++ app/src/main/res/layout/fragment_create_profile.xml
@@ -4,7 +4,7 @@
   android:layout_width="match_parent"
   android:layout_height="match_parent"
   xmlns:app="http://schemas.android.com/apk/res-auto"
-  tools:context=".CreateProfileFragment">
+  tools:context=".ui.CreateProfileFragment">
 
   <com.google.android.material.appbar.AppBarLayout
     android:layout_width="match_parent"
